+++ src/ship.rs
@@ -64,6 +64,11 @@ fn setup(mut commands: Commands, sprites: Res<SpriteAssets>) {
         },
         SpriteBundle {
             texture: sprites.player_ship.clone(),
+            transform: Transform {
+                translation: Vec3::new(0., 0., 100.0),
+                scale: Vec3::splat(0.5),
+                ..default()
+            },
             ..default()
         },
         RigidBody::Dynamic,