+++ src/star.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::*;
+
+#[derive(Component, Clone, Debug)]
+pub struct Star {}